valorank_root/requirements.txt
Django==4.1.4
django-versatileimagefield==2.2
Pillow==9.3.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
django-ckeditor==6.5.1
gunicorn==20.1.0
